# services/gateway/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_fieldstone_fieldstone_services_permits_handlers.listResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      permits:
        items:
          $ref: '#/definitions/handlers.PermitResponse'
        type: array
      total:
        type: integer
    type: object
  github_com_fieldstone_fieldstone_services_permits_handlers.updateStatusRequest:
    properties:
      role:
        type: string
      status:
        type: string
    type: object
...
  handlers.InspectionResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      inspector_id:
        type: string
      notes:
        type: string
      permit_id:
        type: string
      result:
        type: string
      scheduled_at:
        type: string
    type: object
  handlers.PermitDetailResponse:
    properties:
      applicant:
        items:
          type: integer
        type: array
      created_at:
        type: string
      department_id:
        type: string
      expires_at:
        type: string
      id:
        type: string
      inspections:
        items:
          $ref: '#/definitions/handlers.InspectionResponse'
        type: array
      issued_at:
        type: string
      metadata:
        items:
          type: integer
        type: array
      permit_type:
        type: string
      property_address:
        type: string
      status:
        type: string
      submitted_at:
        type: string
      updated_at:
        type: string
    type: object
  handlers.PermitResponse:
    properties:
      applicant:
        items:
          type: integer
        type: array
      created_at:
        type: string
      department_id:
        type: string
      expires_at:
        type: string
      id:
        type: string
      issued_at:
        type: string
      metadata:
        items:
          type: integer
        type: array
      permit_type:
        type: string
      property_address:
        type: string
      status:
        type: string
      submitted_at:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.createPermitRequest:
    properties:
      applicant:
        items:
          type: integer
        type: array
      department_id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      permit_type:
        type: string
      property_address:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Fieldstone Contributors
    url: https://github.com/fieldstone/fieldstone
  description: |-
    A self-hostable platform for city governments to run common civic services.

    Staff endpoints require a Bearer JWT from the configured OIDC provider.
    Citizen endpoints (marked "public") do not require authentication.

    **Development**: set `DEV_DISABLE_AUTH=true` in .env to bypass JWT
    validation locally. Remove before production deployment.
  license:
    name: MIT
    url: https://github.com/fieldstone/fieldstone/blob/main/LICENSE
  title: Fieldstone Civic Services API
  version: 0.1.0
paths:
...
  /v1/permits:
    get:
      parameters:
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Filter by status (submitted, under_review, approved, rejected,
          expired)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fieldstone_fieldstone_services_permits_handlers.listResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List permits
      tags:
      - permits
    post:
      consumes:
      - application/json
      parameters:
      - description: Permit application
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createPermitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PermitResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Metadata validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a permit application
      tags:
      - permits
  /v1/permits/{id}:
    get:
      parameters:
      - description: Permit UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PermitDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a permit with its inspections
      tags:
      - permits
...
  /v1/permits/{id}/status:
    patch:
      consumes:
      - application/json
      description: |-
        Validates the transition via the workflow service. Setting status to
        'approved' also records issued_at. Publishes permit.status_changed.
      parameters:
      - description: Permit UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status transition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fieldstone_fieldstone_services_permits_handlers.updateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PermitResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Transition not allowed by workflow
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update permit status
      tags:
      - permits
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"